order_service/requirements.txt
Django==3.0.4
django-mysql==3.3.0
mysqlclient~=1.4.4
djangorestframework==3.11.0
requests==2.23.0
celery==4.2.1
redis==3.4.1
pika==1.1.0
coverage~=4.5.3
pylint-django
